Make CandidateList available through aecdata module.
