+++ dividebatur/aecdata/__init__.py
@@ -1,7 +1,10 @@
 import itertools
 import csv
 import re
-from collections import namedtuple, defaultdict
+from collections import defaultdict
+
+from .candidatelist import CandidateList
+from .utils import named_tuple_iter, ticket_sort_key
 from ..counter import Ticket
 
 
@@ -12,25 +15,6 @@ def int_or_none(s):
         return int(s)
     except ValueError:
         return None
-
-
-def ticket_sort_key(ticket):
-    "sort key for an ATL ticket, eg. A..Z, AA..ZZ"
-    return (len(ticket), ticket)
-
-
-def named_tuple_iter(name, reader, header, **kwargs):
-    field_names = [t for t in [t.strip().replace('-', '_')
-                               for t in header] if t]
-    typ = namedtuple(name, field_names)
-    mappings = []
-    for field_name in kwargs:
-        idx = field_names.index(field_name)
-        mappings.append((idx, kwargs[field_name]))
-    for row in reader:
-        for idx, map_fn in mappings:
-            row[idx] = map_fn(row[idx])
-        yield typ(*row)
 
 
 class AllCandidates:
